fix windows build of bitfield.hpp

+++ include/libtorrent/bitfield.hpp
@@ -236,7 +236,7 @@ namespace libtorrent
 #ifdef __GNUC__
 					return i * 32 + __builtin_clz(v);
 #else
-					long int pos;
+					DWORD pos;
 					_BitScanReverse(&pos, v);
 					return i * 32 + 31 - pos;
 #endif
@@ -282,7 +282,7 @@ namespace libtorrent
 #ifdef __GNUC__
 					return i * 32 + 31 - __builtin_ctz(last);
 #else
-					long int pos;
+					DWORD pos;
 					_BitScanForward(&pos, last);
 					return i * 32 + 31 - pos;
 #endif
@@ -312,7 +312,7 @@ namespace libtorrent
 #ifdef __GNUC__
 					return i * 32 + 31 - __builtin_ctz(v);
 #else
-					long int pos;
+					DWORD pos;
 					_BitScanForward(&pos, v);
 					return i * 32 + 31 - pos;
 #endif
